Add a tag generator.
Add way of getting a generator of FLV tags, instead of forcing the
user to parse all tags at once. Makes the output appear faster and
makes something appear even if there are subsequent errors in the file.

This is only used by debug-flv, because index-flv needs repeated access
to all tags anyway.

Changed the FLV object API a bit, naming the methods in a bit more
precise way.