Do not fix the index in the metadata in retimestamp-flv.
The retimestamp-flv utility should not touch anything but the
timestamps of the FLV tags. If the index gets out of date due to
retimestamping, the file should be reindexed.

In passing change the atomic retimestamping behaviour to copy the
whole file to a temporary location and retimestamp it in place, which
hopefully will perform better than doing lots of small writes when
copying tag by tag.