Fix handling of date objects.
The spec says that time offsets should be ignored and that the date
object always is expressed in milliseconds since epoch. Some encoders
put garbage in the time offset field and trying to interptet it per
the spec (as offset from GMT in minutes) was leading to errors. A
simple fix is to just ignore it and always produce Python datetime
objects in UTC when parsing AMF dates.

Per bug report from Bobby Impollonia.